requirements.txt
# MLB Automation System - Python Dependencies
# Core dependencies for complete automation pipeline

# Database connectivity
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Data analysis and machine learning
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Graphics and visualization
matplotlib>=3.7.0
seaborn>=0.12.0
Pillow>=10.0.0

# Social media integration
tweepy>=4.14.0

# Web requests and APIs
requests>=2.31.0
urllib3>=2.0.0

# System monitoring and performance
psutil>=5.9.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Date and time handling
python-dateutil>=2.8.0

# Configuration and data formats
PyYAML>=6.0.0
toml>=0.10.0

# Baseball-specific packages
pybaseball>=2.2.0
MLB-StatsAPI>=1.6.0

# Flask for microservices
flask>=2.3.0
flask-cors>=4.0.0
gunicorn>=21.0.0

# Scheduling
apscheduler>=3.10.0

# Email functionality (for monitoring alerts)
secure-smtplib>=0.1.1

# Optional: Advanced ML packages
# xgboost>=1.7.0
# lightgbm>=4.0.0

# Optional: Advanced visualization
# plotly>=5.15.0
# bokeh>=3.2.0